Rename sender package to connection
