+++ src/main/java/com/teragrep/cfe_16/EventManager.java
@@ -58,8 +58,8 @@ import com.teragrep.cfe_16.config.Configuration;
 import com.teragrep.cfe_16.exceptionhandling.EventFieldBlankException;
 import com.teragrep.cfe_16.exceptionhandling.EventFieldMissingException;
 import com.teragrep.cfe_16.exceptionhandling.InternalServerErrorException;
-import com.teragrep.cfe_16.sender.AbstractSender;
-import com.teragrep.cfe_16.sender.SenderFactory;
+import com.teragrep.cfe_16.connection.AbstractSender;
+import com.teragrep.cfe_16.connection.SenderFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
