+++ src/main/java/com/teragrep/cfe_16/connection/TcpSender.java
@@ -43,30 +43,49 @@
  * Teragrep, the applicable Commercial License may apply to this file if you as
  * a licensee so wish it.
  */
-package com.teragrep.cfe_16.sender;
+package com.teragrep.cfe_16.connection;
 
+import com.cloudbees.syslog.SyslogMessage;
+import com.cloudbees.syslog.sender.TcpSyslogMessageSender;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 
-public class SenderFactory {
+public class TcpSender extends AbstractSender {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(SenderFactory.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(TcpSender.class);
+    private TcpSyslogMessageSender sender;
 
-    public static AbstractSender createSender(String type, String hostname, int port) throws IOException {
-        LOGGER.debug("Creating sender for type <[{}]> to <[{}]>:<[{}]>", type, hostname, port);
-        if (type.equalsIgnoreCase("UDP")) {
-            return new UdpSender(hostname, port);
+    public TcpSender(String hostname, int port) {
+        super(hostname, port);
+        this.sender = new TcpSyslogMessageSender();
+        this.sender.setSyslogServerHostname(this.hostname);
+        this.sender.setSyslogServerPort(this.port);
+    }
+
+    @Override
+    public void sendMessages(SyslogMessage[] syslogMessages) throws IOException {
+        LOGGER.debug("Sending messages");
+        for (SyslogMessage syslogMessage : syslogMessages) {
+            this.sender.sendMessage(syslogMessage);
         }
-        else if (type.equalsIgnoreCase("TCP")) {
-            return new TcpSender(hostname, port);
-        }
-        else if (type.equalsIgnoreCase("RELP")) {
-            return new RelpSender(hostname, port);
-        }
-        else {
-            throw new IOException("Invalid sender type: " + type);
-        }
+    }
+
+    @Override
+    public void sendMessage(SyslogMessage syslogMessage) throws IOException {
+        LOGGER.debug("Sending message");
+        this.sender.sendMessage(syslogMessage);
+    }
+
+    @Override
+    public void close() throws IOException {
+        LOGGER.debug("Closing sender");
+        this.sender.close();
+    }
+
+    public void setSsl(boolean ssl) {
+        LOGGER.debug("Set Ssl to <{}>", ssl);
+        this.sender.setSsl(ssl);
     }
 }
